refactor modules to have intuitive structure
