+++ server/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 
 from server.config.factory import settings
-from server.models.managers import create_db_and_tables, ping_redis_server, pool_database_clients
+from server.database.managers import create_db_and_tables, ping_redis_server, pool_database_clients
 from server.routes.auth import router as auth_router
 from server.schemas.base import HealthResponseSchema
